Convert the core to RV64IMAC_Zicsr -- this is a base-width change, not an extension
XLEN 32 -> 64 touches nearly every file, unlike C/Zicsr which built on
top of a stable RV32 base:

- Registers::x[32]/pc/csr[] are now uint64_t. HistoryEntry.pc too.
- Memory's address parameters widen to uint64_t; added read64/write64
  for LD/SD/AMO*.D/LR.D/SC.D. read32's fast path is unaffected (still
  a plain memcpy, alignment-agnostic).
- The ELF loader now parses Elf64_Ehdr/Elf64_Phdr -- a genuinely
  different struct layout from Elf32, not just wider fields (p_flags
  moves position in Phdr). This makes the old ELF32 guest binary
  fundamentally unreadable by this loader -- expected, see the
  companion guest-toolchain commit.
- DecodedInstruction.imm is int64_t; the existing 32-bit immediate
  computations in decode() didn't need touching -- assigning an
  already-sign-extended int32_t into an int64_t field sign-extends
  again for free, which happens to be exactly right for RV64 LUI/
  AUIPC/etc semantics too.
- New opcode families: OP-IMM-32/OP-32 (0x1b/0x3b) for ADDIW/SLLIW/
  SRLIW/SRAIW and ADDW/SUBW/SLLW/SRLW/SRAW/MULW/DIVW/DIVUW/REMW/REMUW
  -- all 32-bit ops whose result sign-extends to 64 (DecodedInstruction
  gained a word_op flag exec_32I/exec_32M check for this). LD/LWU/SD
  added to the existing Load/Store opcodes. AMO ops gained a func3-
  selected width (op_64 flag) covering LR.D/SC.D/AMO*.D alongside the
  existing .W forms.
- exec_32M's MULH/MULHSU/MULHU now need the upper 64 bits of a 128-bit
  product -- uses __int128/unsigned __int128 (a well-supported GCC/
  Clang extension, not standard C++, but this project already builds
  exclusively with g++).
- decode_compressed() rewritten for RV64 semantics specifically:
  several encodings mean something different than on RV32 -- quadrant-
  1 func3=001 is C.ADDIW here (was C.JAL), quadrant-0/2 gained C.LD/
  C.SD/C.LDSP/C.SDSP in the slots RV32 reserved for float loads/stores,
  and quadrant-1's func3=100 func1=1 slot (previously always illegal,
  "RV64-only") now decodes C.SUBW/C.ADDW.
- Dashboard: register grid shows the low 32 bits (16-hex-digit values
  don't fit the two-column layout without overlapping columns), but
  CURR PC and the trace log -- single-width lines, not column-
  constrained -- show the full 64-bit address.

Deliberately RV64-only, not a runtime/compile-time RV32-or-RV64
toggle: supporting both would mean two parallel meanings for every
compressed encoding plus dual immediate-width paths throughout the
decoder, for a core that will only ever run one target in practice.

Verified in two stages. First, zero regression: an Elf64 loader can't
read the old ELF32 binary at all, so this is a hard cutover -- see the
companion commit for the guest rebuild. Second, ran the rebuilt RV64
Doom binary live: correct rendering/HUD/gameplay across multiple
sessions, with the trace log showing genuine RV64-only instructions
executing mid-game (C.ADDIW, LD, SRLIW), no crash.log.

+++ src/debugger.hpp
@@ -9,12 +9,12 @@ public:
 	Debugger();
 
 	void load_breakpoints(const char *path);
-	bool should_halt(uint32_t pc, bool instr_was_illegal);
+	bool should_halt(uint64_t pc, bool instr_was_illegal);
 
 	void dump_log(const Registers &regs, const char *path);
 
 	bool halted;
 
 private:
-	std::vector<uint32_t> breakpoints;
+	std::vector<uint64_t> breakpoints;
 };
